+++ src/Resources/Resources.Test/ScenarioTests/ProviderListBuilder.cs
@@ -90,22 +90,22 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
         {
             public static IList<string> DefaultApiVersions { get; } = new List<string> { "2018-01-01", "2016-01-01" };
             private string Name { get; }
-            private List<AliasPathType> Paths { get; } = new List<AliasPathType>();
+            private List<AliasPath> Paths { get; } = new List<AliasPath>();
 
             public AliasBuilder(string name)
             {
                 this.Name = name;
             }
 
-            public AliasType Alias => new AliasType
+            public Alias Alias => new Alias
             {
                 Name = this.Name,
                 Paths = this.Paths
             };
 
-            public AliasPathType AddAliasPath(string path, IEnumerable<string> apiVersions = null)
+            public AliasPath AddAliasPath(string path, IEnumerable<string> apiVersions = null)
             {
-                var rv = new AliasPathType
+                var rv = new AliasPath
                 {
                     Path = path,
                     ApiVersions = apiVersions?.ToList() ?? AliasBuilder.DefaultApiVersions