+++ src/Synapse/Synapse/Commands/DataPlaneCommands/SparkSession/GetAzureSynapseSparkSession.cs
@@ -9,7 +9,7 @@ using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Synapse
 {
-    [Cmdlet(VerbsCommon.Get, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.SparkSession)]
+    [Cmdlet(VerbsCommon.Get, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.SparkSession, DefaultParameterSetName = GetByNameParameterSet)]
     [OutputType(typeof(PSSynapseSparkSession))]
     public class GetAzureSynapseSparkSession : SynapseCmdletBase
     {