+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -19,7 +19,9 @@
 -->
 ## Upcoming Release
 * Azure Site Recovery added support for protecting proximity placement group virtual machines for Azure to Azure provider.
+* Azure Site Recovery added support for zone to zone replication.
 * Azure Backup Added Long term retention support for Azure FileShare Recovery Points.
+* Azure Backup Added disk exclusion properties to Get-AzRecoveryServicesBackupItem Cmdlet Output.
 * Added private endpoint for Vault credential file for site recovery service.
 
 ## Version 2.8.0